Move functionality to app.py
